Add a missing virtual destructor, fix a warning

+++ ext/native/thin3d/thin3d.cpp
@@ -409,7 +409,7 @@ void ConvertFromRGBA8888(uint8_t *dst, const uint8_t *src, uint32_t dstStride, u
 		case Draw::DataFormat::R8G8B8A8_UNORM:
 		case Draw::DataFormat::UNDEFINED:
 		default:
-			WARN_LOG_REPORT_ONCE(convFromRGBA, G3D, "Unable to convert from format: %d", format);
+			WARN_LOG_REPORT_ONCE(convFromRGBA, G3D, "Unable to convert from format: %d", (int)format);
 			break;
 		}
 	}
